+++ callback.php
@@ -19,13 +19,24 @@ if (!defined('_PS_VERSION_')) {
 
 use Tecnoacquisti\SearchConsole\GscOAuthHandler;
 
+function tecGscAppendCallbackStatus($url, $parameter)
+{
+    if ($url === '') {
+        $url = Tools::getShopDomainSsl(true, true) . __PS_BASE_URI__;
+    }
+
+    return $url . (strpos($url, '?') === false ? '?' : '&') . ltrim($parameter, '&?');
+}
+
 $module = Module::getInstanceByName('tec_searchconsole');
 if ($module && method_exists($module, 'loadModuleAutoloader')) {
     $module->loadModuleAutoloader();
 }
 
-$context = isset($module->context) ? $module->context : null;
-$idShop = isset($context->shop->id) ? (int) $context->shop->id : 1;
+$idShop = (int) Configuration::get('PS_SHOP_DEFAULT');
+if ($idShop <= 0) {
+    $idShop = 1;
+}
 $state = (string) Tools::getValue('state');
 $code = (string) Tools::getValue('code');
 $matchedShopId = GscOAuthHandler::findShopIdByState($state);
@@ -33,19 +44,16 @@ if ($matchedShopId > 0) {
     $idShop = $matchedShopId;
 }
 $adminUrl = (string) Configuration::get('TEC_GSC_ADMIN_RETURN_URL_' . (int) $idShop);
-if ($adminUrl === '') {
-    $adminUrl = isset($context->link) ? $context->link->getAdminLink('AdminTecGsc') : '';
-}
 
 if ($matchedShopId <= 0) {
-    Tools::redirectAdmin($adminUrl . '&gsc_error=csrf');
+    Tools::redirectAdmin(tecGscAppendCallbackStatus($adminUrl, 'gsc_error=csrf'));
     exit;
 }
 
 if ($code === '') {
     GscOAuthHandler::clearState($idShop);
     Configuration::deleteByName('TEC_GSC_ADMIN_RETURN_URL_' . (int) $idShop);
-    Tools::redirectAdmin($adminUrl . '&gsc_error=no_code');
+    Tools::redirectAdmin(tecGscAppendCallbackStatus($adminUrl, 'gsc_error=no_code'));
     exit;
 }
 
@@ -59,5 +67,5 @@ try {
 
 GscOAuthHandler::clearState($idShop);
 Configuration::deleteByName('TEC_GSC_ADMIN_RETURN_URL_' . (int) $idShop);
-Tools::redirectAdmin($adminUrl . ($success ? '&gsc_success=connected' : '&gsc_error=token_failed'));
+Tools::redirectAdmin(tecGscAppendCallbackStatus($adminUrl, $success ? 'gsc_success=connected' : 'gsc_error=token_failed'));
 exit;